Update patch set 4
Patch Set 4:

Biju - I think this was originally your code.  This change would update the ApplicationConfig.debugProperties method to write properties to a log file instead of stdout.  I wanted to make sure that's not going to break anything (as in, there's nothing reading stdout that is expecting to read those properties).

Patch-set: 4