Update patch set 1
Patch Set 1:

This cannot be merged due to merge conflict - please resolve.
Also, please update the commit message so that the abstract (first line) is unique. If there are a series of commits that should be committed in a specific order, please indicate that in abstract (e.g. Add netconf module - 1 of 3)

Patch-set: 1
Label: Verified=0