Update patch set 2
Patch Set 2:

OK - instead of odlparent-lite, you might want to look at the 2 springboot parent poms (for springboot 1 vs 2) instead.  Or, just use oparent, if that works.

The problem was that the parent used by that module wasn't in the directory above - so default relativePath doesn't find it.  When doing daily builds, the <version>s are already reset to the release version (based on version.properties), which don't exist yet in the 'public' nexus (only in staging) ... so without relativePath set properly, compile fails due to missing pom reference

Patch-set: 2