*WIP*: draft of a blueprint for configuring a PNF
This is a still-in-development draft of a blueprint, intended to be
used to configure PNFs. It has been put in review as requested, to make
easier for the community to make comments on it.

The design considers that a blueprint will be created for each possible
PNF.

In some files, there are questions about the structure or content. The-
se questions are surrounded by HTML comment tags <!-- -->, even if the
file itself is not HTML - the intent is just to clearly highlight the
questions. All of them are preceded by the string "QUESTION", to facil-
itate search.

Change-Id: Ib3ed2afb80693429dc23b7ac73c5cbe49b3f5c9a
Issue-ID: CCSDK-1107
Signed-off-by: Rodrigo Ottero <rodrigo.ottero@est.tech>